+++ examples/src/polytype.mli
@@ -1,3 +1,5 @@
 type 'a llist =
   | LCons of 'a * (unit -> 'a llist)
   | LNil
+
+type box = Box : 'a -> box